requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
h5py==3.8.0
idna==3.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.5
packaging==23.1
Pillow==9.2.0
pip==23.1.2
protobuf==3.20.3
PyMCubes==0.1.4
PySocks==1.7.1
requests==2.31.0
scipy==1.7.3
setuptools==67.7.2
six==1.16.0
tensorboardX==2.6
torch==1.12.1+cu102
torchaudio==0.12.1
torchvision==0.13.1+cu102
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
wheel==0.40.0
